move api and util to /lib/
