+++ examples/Web/web/src/components/Browse/Browse.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import api from '../../api';
+import api from '../../lib/api';
 
 import './Browse.css';
