+++ examples/Web/web/src/components/Shared/FileList.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { formatSeconds, formatBytes, getFileName } from '../../util';
+import { formatSeconds, formatBytes, getFileName } from '../../lib/util';
 
 import { 
     Header,
