+++ examples/Web/web/src/components/Transfers/TransferList.js
@@ -4,7 +4,7 @@ import {
     Checkbox
 } from 'semantic-ui-react';
 
-import { formatBytes, getFileName } from '../../util';
+import { formatBytes, getFileName } from '../../lib/util';
 
 import { 
     Header,
