+++ examples/Web/web/src/lib/util.js
